// src/api/generated/zod.ts
export const createGoalApiV1GoalsPostBodyTitleMax = 255;

export const createGoalApiV1GoalsPostBodyTargetValueRegExpTwo = new RegExp('^(?!^[-+.]\*$)[+-]?0\*\\d\*\\.?\\d\*$');
export const createGoalApiV1GoalsPostBodyCurrencyMaxOne = 16;



export const createGoalApiV1GoalsPostBody = zod.object({
  "type": zod.enum(['target_net_worth', 'target_return', 'target_allocation', 'custom']),
  "title": zod.string().min(1).max(createGoalApiV1GoalsPostBodyTitleMax),
  "target_value": zod.union([zod.number(),zod.string().regex(createGoalApiV1GoalsPostBodyTargetValueRegExpTwo),zod.null()]).optional(),
  "currency": zod.union([zod.string().max(createGoalApiV1GoalsPostBodyCurrencyMaxOne),zod.null()]).optional(),
  "target_allocation": zod.union([zod.record(zod.string(), zod.unknown()),zod.null()]).optional(),
  "target_date": zod.union([zod.string().date(),zod.null()]).optional(),
  "status": zod.union([zod.enum(['active', 'achieved', 'archived']),zod.null()]).optional()
})