// src/api/generated/schema.ts
export interface components {
    schemas: {
// ...
        GoalOut: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            type: components["schemas"]["GoalType"];
            /** Title */
            title: string;
            /** Target Value */
            target_value?: string | null;
            /** Currency */
            currency?: string | null;
            /** Target Allocation */
            target_allocation?: {
                [key: string]: unknown;
            } | null;
            /** Target Date */
            target_date?: string | null;
            status: components["schemas"]["GoalStatus"];
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}